refactor: 파일 구조 개선 파일명 변경
- page와 section의 경우 명칭에 명시
- 특정 페이지 구성 컴포넌트를 page 하위 폴더로 이동
- 페이지의 page & section을 제외한 하위 컴포넌트를
  페이지별 components 폴더 내에 구성
- 공통 컴포넌트를 components 내에 두고,
  layout 컴포넌트의 경우 별도로 하위 폴더 구성
- convertToMoney 대신 toLocaleString 사용
- 불필요한 App.js 제거
