+++ src/pages/itemListPage/components/Item.js
@@ -1,7 +1,6 @@
 import { Link } from "react-router-dom";
 
-import { convertToMoney } from "../../utils/util.js";
-import heartIcon from "../../assets/heart_inactive.svg";
+import heartIcon from "../../../assets/heart_inactive.svg";
 import "./Item.css";
 
 export default function Item({ data }) {
@@ -11,7 +10,7 @@ export default function Item({ data }) {
       <img className="item__image" src={images[0]} alt={description} />
       <div className="item__content">
         <span className="item__title">{name}</span>
-        <span className="item__price">{convertToMoney(price)}</span>
+        <span className="item__price">{price.toLocaleString() + "원"}</span>
         <div className="item__favorite">
           <img className="item__favoriteIcon" src={heartIcon} alt="좋아요 수" />
           <span>{favoriteCount}</span>
