--- src/pages/secondhand/ItemListPage.js
@@ -1,13 +1,0 @@
-import BestItemList from "../../components/secondhand/BestItemList";
-import AllItemList from "../../components/secondhand/AllItemList";
-
-import "./ItemListPage.css";
-
-export default function ItemListPage() {
-  return (
-    <main className="main">
-      <BestItemList />
-      <AllItemList />
-    </main>
-  );
-}
