--- src/utils/util.js
@@ -1,3 +1,0 @@
-export function convertToMoney(num) {
-  return num.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",") + "원";
-}
